+++ src/library/Uno.Material/Styles/Application/Common/Converters.xaml
@@ -1,54 +1,55 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 					xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-					xmlns:um="using:Uno.Material">
+					xmlns:ut="using:Uno.Themes">
 
-	<um:FromBoolToValueConverter x:Key="MaterialTrueToVisible"
+	<ut:FromBoolToValueConverter x:Key="MaterialTrueToVisible"
 								 TrueValue="Visible"
 								 FalseValue="Collapsed"
 								 NullValue="Collapsed" />
 
-	<um:FromBoolToValueConverter x:Key="MaterialTrueToCollapsed"
+	<ut:FromBoolToValueConverter x:Key="MaterialTrueToCollapsed"
 								 TrueValue="Collapsed"
 								 FalseValue="Visible"
 								 NullValue="Visible" />
 
-	<um:FromEmptyStringToValueConverter x:Key="MaterialEmptyToFalse"
+	<ut:FromEmptyStringToValueConverter x:Key="MaterialEmptyToFalse"
 										NotNullOrEmptyValue="True"
 										NullOrEmptyValue="False" />
 
-	<um:FromEmptyStringToValueConverter x:Key="MaterialEmptyToTrue"
+	<ut:FromEmptyStringToValueConverter x:Key="MaterialEmptyToTrue"
 										NotNullOrEmptyValue="False"
 										NullOrEmptyValue="True" />
 
-	<um:FromEmptyStringToValueConverter x:Key="MaterialEmptyToVisible"
+	<ut:FromEmptyStringToValueConverter x:Key="MaterialEmptyToVisible"
 										NotNullOrEmptyValue="Collapsed"
 										NullOrEmptyValue="Visible" />
 
-	<um:FromEmptyStringToValueConverter x:Key="MaterialEmptyToCollapsed"
+	<ut:FromEmptyStringToValueConverter x:Key="MaterialEmptyToCollapsed"
 										NotNullOrEmptyValue="Visible"
 										NullOrEmptyValue="Collapsed" />
 
-	<um:FromNullToValueConverter x:Key="MaterialNullToCollapsedConverter"
+	<ut:FromNullToValueConverter x:Key="MaterialNullToCollapsedConverter"
 								 NotNullValue="Visible"
 								 NullValue="Collapsed" />
 
-	<um:FromNullToValueConverter x:Key="MaterialNullToVisibleConverter"
+	<ut:FromNullToValueConverter x:Key="MaterialNullToVisibleConverter"
 								 NotNullValue="Visible"
 								 NullValue="Collapsed" />
 
-	<um:FromNullToValueConverter x:Key="MaterialNullToTransparent"
+	<ut:FromNullToValueConverter x:Key="MaterialNullToTransparent"
 								 NotNullValue="1"
 								 NullValue="0" />
 	
-	<um:FromEmptyStringOrNullObjectToValueConverter x:Key="MaterialEmptyOrNullToVisible"
+	<ut:FromEmptyStringOrNullObjectToValueConverter x:Key="MaterialEmptyOrNullToVisible"
 													NotEmptyOrNullValue="Collapsed"
 													EmptyOrNullValue="Visible" />
 
-	<um:FromEmptyStringOrNullObjectToValueConverter x:Key="MaterialEmptyOrNullToCollapsed"
+	<ut:FromEmptyStringOrNullObjectToValueConverter x:Key="MaterialEmptyOrNullToCollapsed"
 													NotEmptyOrNullValue="Visible"
 													EmptyOrNullValue="Collapsed" />
 
-	<um:StringFormatConverter x:Key="StringFormatConverter" />
-	<um:FirstCharacterConverter x:Key="FirstCharacterConverter" />
+	<ut:StringFormatConverter x:Key="StringFormatConverter" />
+	<ut:FirstCharacterConverter x:Key="FirstCharacterConverter" />
+	<ut:ToUpperConverter x:Key="ToUpperConverter"/>
 
 </ResourceDictionary>
