+++ src/library/Uno.Material/Styles/Controls/v1/Button.xaml
@@ -4,11 +4,12 @@
 					xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 					xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 					xmlns:um="using:Uno.Material"
+					xmlns:ut="using:Uno.Themes"
 					xmlns:not_win="http://uno.ui/not_win"
 					mc:Ignorable="d not_win">
 
 	<!-- Documented variables -->
-	<um:ToUpperConverter x:Key="ToUpperConverter" />
+	<!--<ut:ToUpperConverter x:Key="ToUpperConverter" />-->
 	<Thickness x:Key="MaterialButtonPadding">16,8</Thickness>
 	<CornerRadius x:Key="MaterialButtonCornerRadius">4</CornerRadius>
 	<x:Double x:Key="MaterialButtonFontSize">14</x:Double>
