+++ src/library/Uno.Material/Styles/Controls/v1/TextBox.xaml
@@ -5,10 +5,11 @@
 					xmlns:not_macos="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 					xmlns:macos="http://uno.ui/macos"
 					xmlns:um="using:Uno.Material"
+					xmlns:ut="using:Uno.Themes"
 					mc:Ignorable="d macos">
 
 	<!-- Converters -->
-	<um:FromEmptyStringToValueConverter x:Key="EmptyToCompositeTransformTranslateY"
+	<ut:FromEmptyStringToValueConverter x:Key="EmptyToCompositeTransformTranslateY"
 										NotNullOrEmptyValue="7"
 										NullOrEmptyValue="0" />
