+++ src/library/Uno.Material/Styles/Controls/v2/Button.xaml
@@ -4,12 +4,13 @@
 					xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 					xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 					xmlns:um="using:Uno.Material"
+					xmlns:ut="using:Uno.Themes"
 					xmlns:toolkit="using:Uno.UI.Toolkit"
 					xmlns:not_win="http://uno.ui/not_win"
 					mc:Ignorable="d not_win">
 
 	<!-- Converters -->
-	<um:FromNullToValueConverter x:Key="NullToTextButtonMarginConverter"
+	<ut:FromNullToValueConverter x:Key="NullToTextButtonMarginConverter"
 								 NotNullValue="0,0,2,0"
 								 NullValue="0" />
