+++ src/library/Uno.Material/Styles/Controls/v2/RadioButton.xaml
@@ -2,6 +2,7 @@
 					xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 					xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 					xmlns:um="using:Uno.Material"
+					xmlns:ut="using:Uno.Themes"
 					xmlns:android="http://uno.ui/android"
 					xmlns:ios="http://uno.ui/ios"
 					xmlns:wasm="http://uno.ui/wasm"
@@ -131,12 +132,12 @@
 
 
 	<!--#region Converters-->
-	<um:FromBoolToValueConverter x:Key="RadioButtonIsCheckedStateCircleVisible"
+	<ut:FromBoolToValueConverter x:Key="RadioButtonIsCheckedStateCircleVisible"
 								 TrueValue="Visible"
 								 FalseValue="Collapsed"
 								 NullValue="Collapsed" />
 
-	<um:FromBoolToValueConverter x:Key="RadioButtonIsCheckedToStateCircleCollapsed"
+	<ut:FromBoolToValueConverter x:Key="RadioButtonIsCheckedToStateCircleCollapsed"
 								 TrueValue="Collapsed"
 								 FalseValue="Visible"
 								 NullValue="Collapsed" />
