+++ src/library/Uno.Themes/Converters/ToUpperConverter.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+#if WinUI
+using Microsoft.UI.Xaml.Data;
+#else
+using Windows.UI.Xaml.Data;
+#endif
+
+namespace Uno.Themes;
+
+public class ToUpperConverter : IValueConverter
+{
+	public object Convert(object value, Type targetType, object parameter, string language)
+	{
+		return value is string text ? text.ToUpperInvariant() : value;
+	}
+
+	public object ConvertBack(object value, Type targetType, object parameter, string language)
+	{
+		return null;
+	}
+}
